backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Docker for FastAPI backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.134.0",
    "numpy>=2.4.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pygeohash>=1.2.0",
    "sqlalchemy>=2.0.47",
    "uvicorn[standard]>=0.41.0",
    "python-jose[cryptography]>=3.3.0", #this one is need to verify JWT tokens
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]  # Enable pycodestyle (E), Pyflakes (F), and isort (I)
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"  # Explicitly set the loop scope
